Consolidate ownership to outwriter
